feat: Add sendTaskLastHeartbeat
sendTaskHeartbeat updates an execution's context. One state machine can call sendTaskHeartbeat on
another to send period updates. However, if it calls sendTaskSuccess to indicate that no further
updates are forthcoming and that receiver can advance to the next state, any output sent will
_overwrite_ rather than update the receiving execution's context. This could be really quite bad.
The new sendTaskLastHeartbeat allows a state machine to send an update and also indicate no further
updates are coming. The receiving state machine is updated, and then advances to the next state.
This is a niche use case, but when you need it, you really need it.