fix: It was possible for parallel and map parents states to complete more than once. This would be b
I feel like the solution to this has been in front of us all the time, and we've never been able to
see it. Run each branch with sendResponse: 'COMPLETE', then just wait on them all. When that fires,
all the children must have completed, so we can check to see if they all succeeded or if any failed,
and proceed accordingly.